# internal/api/apimodel/model_parse_result.go
/*
 * Set and Tune API
 *
 * API for managing sets and tunes
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apimodel

type ParseResult struct {
	Success bool `json:"success,omitempty"`

	Message string `json:"message,omitempty"`
}